Refactors MapManager to not do any networking. Now, ServerMapManager and ClientMapManager handle any networking.
