+++ Robust.Shared/Map/MapManager.cs
@@ -6,15 +6,13 @@ using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Log;
 using Robust.Shared.Maths;
-using Robust.Shared.Physics;
-using Robust.Shared.Physics.Dynamics;
 using Robust.Shared.Timing;
 using Robust.Shared.Utility;
 
 namespace Robust.Shared.Map
 {
     /// <inheritdoc cref="IMapManager"/>
-    internal partial class MapManager : IMapManagerInternal
+    internal class MapManager : IMapManagerInternal
     {
         [Dependency] private readonly IGameTiming _gameTiming = default!;
         [Dependency] private readonly IEntityManager _entityManager = default!;
@@ -38,7 +36,7 @@ namespace Robust.Shared.Map
         ///     so that you don't spam an event for each of the million station tiles.
         /// </summary>
         /// <inheritdoc />
-        public event EventHandler<GridChangedEventArgs>? GridChanged;
+        public virtual event EventHandler<GridChangedEventArgs>? GridChanged;
 
         /// <inheritdoc />
         public event EventHandler<MapEventArgs>? MapCreated;
@@ -52,14 +50,11 @@ namespace Robust.Shared.Map
         private MapId HighestMapID = MapId.Nullspace;
         private GridId HighestGridID = GridId.Invalid;
 
-        private readonly HashSet<MapId> _maps = new();
-        private readonly Dictionary<MapId, GameTick> _mapCreationTick = new();
+        private protected readonly HashSet<MapId> _maps = new();
+        private protected readonly Dictionary<MapId, GameTick> _mapCreationTick = new();
 
-        private readonly Dictionary<GridId, MapGrid> _grids = new();
-        private readonly Dictionary<MapId, EntityUid> _mapEntities = new();
-
-        private readonly List<(GameTick tick, GridId gridId)> _gridDeletionHistory = new();
-        private readonly List<(GameTick tick, MapId mapId)> _mapDeletionHistory = new();
+        private protected readonly Dictionary<GridId, MapGrid> _grids = new();
+        private protected readonly Dictionary<MapId, EntityUid> _mapEntities = new();
 
 #if DEBUG
         private bool _dbgGuardInit = false;
@@ -188,7 +183,7 @@ namespace Robust.Shared.Map
         }
 
         /// <inheritdoc />
-        public void DeleteMap(MapId mapID)
+        public virtual void DeleteMap(MapId mapID)
         {
 #if DEBUG
             DebugTools.Assert(_dbgGuardRunning);
@@ -220,10 +215,6 @@ namespace Robust.Shared.Map
                 _mapEntities.Remove(mapID);
             }
 
-            if (_netManager.IsClient)
-                return;
-
-            _mapDeletionHistory.Add((_gameTiming.CurTick, mapID));
             Logger.InfoS("map", $"Deleting map {mapID}");
         }
 
@@ -577,7 +568,7 @@ namespace Robust.Shared.Map
                 yield return GridId.Invalid;
         }
 
-        public void DeleteGrid(GridId gridID)
+        public virtual void DeleteGrid(GridId gridID)
         {
 #if DEBUG
             DebugTools.Assert(_dbgGuardRunning);
@@ -596,9 +587,6 @@ namespace Robust.Shared.Map
             _grids.Remove(grid.Index);
 
             OnGridRemoved?.Invoke(mapId, gridID);
-
-            if (_netManager.IsServer)
-                _gridDeletionHistory.Add((_gameTiming.CurTick, gridID));
         }
 
         public MapId NextMapId()
